// src/store/api/lib/enums.ts
export enum ApiTags {
  USER = "USER",
  LISTING = "LISTING",
  CURRENT_USER = "CURRENT_USER",
  USER_LISTINGS = "USER_LISTINGS",
  FULL_CATEGORIES_LIST = "FULL_CATEGORIES_LIST",

  USER_SEARCH_REGION_HISTORY = "USER_SEARCH_REGION_HISTORY",

  CURRENT_LISTING = "CURRENT_LISTING",
  VERIFIED_LISTINGS = "VERIFIED_LISTINGS",
  LISTING_CATEGORIES = "LISTING_CATEGORIES",
  LISTING_TYPE_OF_PLACE = "LISTING_TYPE_OF_PLACE",
  VERIFIED_LISTINGS_BY_PARAMS = "VERIFIED_LISTINGS_BY_PARAMS",

  USER_CHATS = "USER_CHATS",
  CURRENT_CHAT = "CURRENT_CHAT",
  USER_NOTIFICATIONS = "USER_NOTIFICATIONS",
}

export enum NotificationTypes {
  MANAGE = "MANAGE",
  GENERAL = "GENERAL",
  INBOX_MESSAGE = "INBOX_MESSAGE",
}